move Shot class to separate file
