+++ models.py
@@ -5,27 +5,6 @@ import pygame as pg
 from settings import Settings
 
 
-class Shot(pg.sprite.Sprite):
-    """a bullet the Player sprite fires."""
-
-    speed = -11
-    images: List[pg.Surface] = []
-
-    def __init__(self, pos, *groups):
-        pg.sprite.Sprite.__init__(self, *groups)
-        self.image = self.images[0]
-        self.rect = self.image.get_rect(midbottom=pos)
-
-    def update(self):
-        """called every time around the game loop.
-
-        Every tick we move the shot upwards.
-        """
-        self.rect.move_ip(0, self.speed)
-        if self.rect.top <= 0:
-            self.kill()
-
-
 class Explosion(pg.sprite.Sprite):
     """An explosion. Hopefully the Alien and not the player!"""
